Добавлен BaseTrackerViewController, в который перенесена общая логика создания событий из NewRegularViewController и NewIrregularViewController. Добавлена проверка ввода пользователем всех данных до создания трекера.
